Allow specifying destination for init
Previously it always initialized the current directory.  With this
change it operates more like `git clone`.  `gg init` now takes an
optional 4th parameter, which is the destination folder.  If that was
not specified, it derives the folder name from the repository.

Once the destination is determined, it will check if it exists.  It may
already be an initialized gg directory or have any files in it (both
fail).  If empty, installation continues as expected.

Thanks to @kyle-long and @uilwen for feedback and suggestions.